# docs/openapi.yaml
basePath: /proxy/v1
definitions:
  common.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  common.PaginatedSearchResponse:
    properties:
      metadata:
        $ref: '#/definitions/common.PaginationMetadata'
      results:
        items:
          $ref: '#/definitions/models.SearchItem'
        type: array
    type: object
  common.PaginationMetadata:
    properties:
      page:
        type: integer
      total_pages:
        type: integer
      total_results:
        type: integer
    type: object
...
  models.Author:
    properties:
      name:
        type: string
      type:
        type: string
    type: object
...
  models.Episode:
    properties:
      description:
        type: string
      duration_minutes:
        type: integer
      episode_number:
        type: integer
      episode_type:
        type: string
      id:
        type: string
      image_url:
        type: string
      release_date:
        type: string
      season_number:
        type: integer
      title:
        type: string
    type: object
...
  models.ImageEntry:
    properties:
      image_url:
        type: string
      size:
        $ref: '#/definitions/models.ImageSize'
      type:
        $ref: '#/definitions/models.ImageType'
    type: object
  models.ImageSize:
    enum:
    - standard
    - original
    type: string
    x-enum-varnames:
    - ImageSizeStandard
    - ImageSizeOriginal
  models.ImageType:
    enum:
    - poster
    - gallery
    type: string
    x-enum-varnames:
    - ImageTypePoster
    - ImageTypeGallery
...
  models.Platform:
    properties:
      image_url:
        type: string
      name:
        type: string
    type: object
...
  models.SearchItem:
    properties:
      authors:
        items:
          $ref: '#/definitions/models.Author'
        type: array
      description:
        type: string
      id:
        type: string
      image_url:
        type: string
      original_title:
        type: string
      release_date:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
  models.SeasonResponse:
    properties:
      description:
        type: string
      episodes:
        items:
          $ref: '#/definitions/models.Episode'
        type: array
      id:
        type: string
      image_url:
        type: string
      images:
        items:
          $ref: '#/definitions/models.ImageEntry'
        type: array
      number_of_episodes:
        type: integer
      platforms:
        additionalProperties:
          items:
            $ref: '#/definitions/models.Platform'
          type: array
        type: object
      release_date:
        type: string
      season_number:
        type: integer
      title:
        type: string
      tv_show_name:
        type: string
      type:
        type: string
    type: object
  models.TVShowResponse:
    properties:
      authors:
        items:
          $ref: '#/definitions/models.Author'
        type: array
      description:
        type: string
      id:
        type: string
      image_url:
        type: string
      images:
        items:
          $ref: '#/definitions/models.ImageEntry'
        type: array
      imdb_id:
        type: string
      number_of_episodes:
        type: integer
      number_of_seasons:
        type: integer
      original_title:
        type: string
      platforms:
        additionalProperties:
          items:
            $ref: '#/definitions/models.Platform'
          type: array
        type: object
      release_date:
        type: string
      seasons:
        items:
          $ref: '#/definitions/models.SeasonResponse'
        type: array
      status:
        type: string
      tagline:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: |-
    REST API proxy that aggregates metadata from TMDB, IGDB, Spotify, and OpenLibrary.
    Authentication: send API key via X-Api-Key header or Authorization: Bearer <token>.
    Rate limit: 60 requests/minute on protected endpoints.
  title: Denn Proxy API
  version: "1.0"
paths:
...
  /tv_shows/{id}:
    get:
      parameters:
      - description: TMDB TV show ID
        in: path
        name: id
        required: true
        type: integer
      - description: ISO 3166-1 alpha-2 country code for platform availability (default
          US)
        in: header
        name: X-User-Country
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TVShowResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - ApiKeyHeader: []
      - BearerAuth: []
      summary: Get TV show details
      tags:
      - TV Shows
  /tv_shows/{id}/seasons/{season_number}:
    get:
      parameters:
      - description: TMDB TV show ID
        in: path
        name: id
        required: true
        type: integer
      - description: Season number (1-based)
        in: path
        name: season_number
        required: true
        type: integer
      - description: ISO 3166-1 alpha-2 country code for platform availability (default
          US)
        in: header
        name: X-User-Country
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SeasonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - ApiKeyHeader: []
      - BearerAuth: []
      summary: Get season details with episodes
      tags:
      - TV Shows
  /tv_shows/bulk:
    get:
      description: Fetches full details for up to 50 TV shows in parallel. Failed
        IDs are silently omitted.
      parameters:
      - description: Comma-separated TMDB TV show IDs (max 50)
        example: 1396,66732
        in: query
        name: ids
        required: true
        type: string
      - description: ISO 3166-1 alpha-2 country code for platform availability (default
          US)
        in: header
        name: X-User-Country
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.TVShowResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - ApiKeyHeader: []
      - BearerAuth: []
      summary: Get multiple TV shows by ID
      tags:
      - TV Shows
  /tv_shows/search:
    get:
      parameters:
      - description: Search term
        in: query
        name: query
        required: true
        type: string
      - default: 1
        description: Page number (min 1)
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Results per page (max 50)
        in: query
        maximum: 50
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.PaginatedSearchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - ApiKeyHeader: []
      - BearerAuth: []
      summary: Search TV shows
      tags:
      - TV Shows
  /tv_shows/trending:
    get:
      parameters:
      - default: 1
        description: Page number (min 1)
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Results per page (max 50)
        in: query
        maximum: 50
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.PaginatedSearchResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - ApiKeyHeader: []
      - BearerAuth: []
      summary: Get popular/trending TV shows
      tags:
      - TV Shows
securityDefinitions:
  ApiKeyHeader:
    description: API key passed in the X-Api-Key header
    in: header
    name: X-Api-Key
    type: apiKey
  BearerAuth:
    description: 'API key passed as a Bearer token: "Authorization: Bearer <key>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"